injective rewrite, bugfixes
previously, when trying to rewrite a goal like:

Goal : ... add(↑a,b) ...

with an equality like:

e : Nat{add(a,b) == X}

this would have no effect, since 'add(↑a,b) != add(a,b)', and the
rewrite function doesn't call whnf - and it can't, as that would result
in looping (inside recursive functions, for example). to solve this,
I've made rewrite aware of native constructors, so that it will whnf and
enter only if there is a head constructor to peel. that solves the
situation above. yet, while trying to prove add-comm, I also noticed
rewriting certain goals resulted in "CantInfer" errors. after debugging,
this was tracked down to the line:

whnfAppLam lv book f x = whnf lv book (f (whnf lv book x))

which took the whnf of arguments even when not needed, removing Ann
nodes that are used by the type-checker. changing it to:

whnfAppLam lv book f x = whnf lv book (f x)

fixes this bug.

this also improves formatting of applications

this also solves a typo on the Ref-Ref equality