introduce parseExpr to avoid operator ambiguity
to avoid confusion in:

foo<x: A>

which could be interpreted as ending in a greater-than operator, we're
introducing a 'parseExpr' parser, which is like 'parseTerm', but without
non-tight operators. this can be used in positions where we expect a
closing symbol that could be confused with an operator.

another place where this is used is in the short def syntax:

def Foo : Nat = 123

(here, 'Nat =' could be interpreted as an assignment)