flatten rework - done
the flatten algorithm has been completely reworked. it is considerably
more robust and general now. it works by deeply flattening the pattern
match into a tree of if-lets, by peeling a a single layer of a single
constructor out:

    match x y:
      case @Z{}      @A: A
      case @S{@Z}    @B: B
      case @S{@S{p}} @C: C(p)
    ------------------------- peel @Z , peel @S{k}
    match x:
      with y
      case @Z: # ↓ peel @Z
        match y:
          case @A: A
      case @S{k}: # ↓ peel @S{k}
        match k y:
          case @Z    @B: B
          case @S{p} @C: C(p)

and then repeating this until every match has exactly:
- 1 scrutinee
- 1 value case
- 1 default case

we then proceed to 'unflatten' a little bit by merging if-lets into
one-layer-deep matches - which is closer to the lambda-match form - and,
then, we finally convert it to lambda matches, completely removing every
Pat constructor from the source term.

there are some interesting ways to change the behavior of the flattener.
for example, we could auto-move constructors, fully linearizing the
term, by replacing the following lines:

flattenPat d pat@(Pat (s:ss) ms ((((cut->Var k i):ps),rhs):cs)) =
  flatten d $ Pat ss ((k,s):ms) (joinVarCol (d+1) k (((Var k i:ps),rhs):cs))
...
flattenPat d pat@(Pat (s:ss) ms cs@((((cut->p):_),_):_)) =
...
    moves   = ms ++ map (\ (s,i) -> (patOf (d+i) s, s)) (zip ss [0..])

the problem with that is that the type checker is not smart enough to
handle all these moved variables; in the VecInd proof, we get errors
like 'Vec(A,n) != Vec(A,n)', because the moved variables are different
than the source variables. I think that strengthening the type checker
to understand these situations, and then making the flattener linearize
everything, is the way to go, as that'd result in better HVM terms and
more specialized goals when proving theorems. it would also let us get
rid of the Core.Move module.

as an alternative, we could let the flattener receive a flag to enable
full linearization, and use it when compiling to HVM.

another interesting flag would be to *not* convert native flat
pattern-matches to lambda-matches - i.e., skipping the last step of the
flattener. this could be useful on the JS target, resulting in more
efficient terms (move lambdas prevent λ-free switch statements, which
are important for JS performance).

another option that is attractive to me is to just remove moves
entirely, and find another way to make the type checker capable of
specializing context variables w.r.t. matched values, and another way to
linearize pattern-matches for HVM compilation.