definition lifting WIP
on BendGen core, by convention, all top-level definitions must be in a
canonical format based on case-trees, where the form `(λ{...} x)`, i.e.,
in, a lambda-match applied to an argument, is not allowed inside the
body. as such, definitions like:

    def G(x: Nat) -> Nat:
      (a,b) = (1n,2n)
      F(a,b,x)

which are desuraged as:

    def G(x: Nat) -> Nat:
      (λ{(,): λa. λb. F(a,b,x)} (1n,2n))

are not allowed, and are CantInfer errors on the core checker.

this is important for multiple reasons:
- it is the fastest format to type-check
- it is the canonical format BendGen searches proofs on
- it is also the fastest format to compute
  - HVM1 used a similar format
  - HVM2 *requires* on it for recursion (due to ultra-strictness)
  - HVM3 might use it? (to avoid large RHS allocations)
- easy to translate to/from Agda-like equational notation
- makes termination on equality much easier (due to guarded deref)
- is needed for theorem proving, or else goals can get hopelessly stuck
  see: https://discord.com/channels/912426566838013994/1387066665933471924/1396901510851334186

yet, while this format is great for computation, not being able to write
`G` (and similar) programs is very inconvenient for the end-user.

in this update, Bend's checker will, rather than erroring, fix these
definitions by lifting them into multiple definitions. so, for example,
the 'G' definition above will become:

    def G(x: Nat) -> Nat:
      (G0 (1n,2n))

    def G0 : Nat -> (Nat & Nat) -> Nat =
      λx. λ{(,): λa. λb. F(a,b,x)}

this is a delicate update that can (and will) break in many ways.
generating terms dynamically like that on the type checker is complex
due to the variable binding structure. many small things can go wrong.
and even if all was right, it is still lacking eta-reductions, so that
the terms we generated by the checker are in the same shape as the terms
that come from the parser.

much cleanup is due, but this is a first step in a needed direction