lets must actually bind vars
I just realized one of the issues with @_paulocavalcanti's proposal of
substituting the value on let checking: if we have something like:

```
F =
  let x = value
  ...
  ... x x ...
```

where x is used more than once, and we detach the `... x x ...` part to
a separate function, we will get:

```
F =
  let x = value
  G(x)

G = λx.
  ... value value ...
```

i.e., the detached fn will have two copies of `value`, because we've
inlined it. so having Let subst like that can't be done

we need another solution to let vars participating of equality