remove detach feature
from Discord:

so, thinking about this issue, since it is a little bit complex to get
right, specially because it requires compile-time synthesis /
modification of book terms, which is very problematic for many reasons
(just an example: if a term "earlier in the file" normalizes an
expression using F, and F is later modified, then the "past reduction"
would be incorrect...). we don't want to mess up here - correctness is a
higher priority than features now - so, I'll just disable this entirely
in the near term. when the user writes a match that isn't allowed, it
will just throw an Unsupported error, asking them to manually split the
definition. this will be annoying, but is the right call for now. when
the repo is much more stable, thoroughly tested / scaled / bug-free, we
can revisit this feature, parting from a more solid foundation than what
we have right now

---

future goal: have a file like

    def is_even(x: Nat) -> Bool:
      match x:
        case 0n:
          True
        case 1n:
          False
        case 2n+p:
          is_even(p)

    def T(n: Nat) -> Set:
      match is_even(n):
        case True:
          Unit
        case False:
          Nat

    def foo(n: Nat) -> T(n):
      match is_even(n):
        case False:
          0n
        case True:
          ()

elaborate / be equivalent to:

  def is_even(x: Nat) -> Bool:
    match x:
      case 0n:
        True
      case 1n:
        False
      case 2n+p:
        is_even(p)

  def T_aux(x: Bool) -> Set:
    match x:
      case True:
        Unit
      case False:
        Nat

  def T(n: Nat) -> Set:
    T_aux(is_even(n))

  def foo_aux(x: Bool) -> T_aux(x):
    match x:
      case False:
        0n
      case True:
        ()

  def foo(n: Nat) -> T(n):
    foo_aux(is_even(n))

without the user needing to manually split it.