Fix nested pattern match error when eta reducing.
Eta reduce algorithm was not recognizing pattern matches where the
lambda order is different then the application order, as described in:

The problem is in the isEtaLong function:
since we do not consider pattern matches on the isEtaLong function, we never cover all the cases because it will "abort" and continue before entering the arms of each match to see if it find's an application of the arg we're inspecting.

For example:

In the term: λs:Char[]. λb:Bool. (λ{False:();True:(λ{[]:();<>:λh. λt. ()})(s)})(b)

call reduceEtas com λs:Char[]. λb:Bool. (λ{False:();True:(λ{[]:();<>:λh. λt. ()})(s)})(b)

Then we call isEtaLong with the inside term (applying Var s to the lambda body):

λb:Bool. (λ{False:();True:(λ{[]:();<>:λh. λt. ()})(s)})(b)
We find intermediate b, so inject it and go deeper
ETALONG: (λ{False:();True:(λ{[]:();<>:λh. λt. ()})(s)})(b)
Since this is an app but the variable is not the target, pass through

Since we fall in a match now, it goes to the _ -> Nothing case and "halts". So we skipped the s lambda without eta reducing with the s application we have inside the match.